requirements.txt
torch
torchvision
opencv-python
mss
stockfish
numpy
Pillow
